007demo/requirements.txt
langchain==1.2.7
langchain-community==0.4.1
langchain-core==1.2.7
langgraph==1.0.7
langchain-openai==1.1.7
mcp==1.25.0
python-dotenv==1.2.1
openai==2.15.0
httpx==0.28.1
beautifulsoup4==4.14.3
fastapi==0.128.0
uvicorn==0.40.0
